test: adiciona mais testes do fornecedor, separados em unitarios e 2e2.
